requirements.txt
django-oauth-toolkit
djangorestframework
django
django-allauth
